Fix some install steps (#154)
Fix the static analysis docs to only make you install
`typescript-eslint` and the DBOS `eslint` plugin. Earlier, for the flat
config setup path, there was a conflict with installing both
`typescript-eslint` and `eslint` manually, but this solves that issue
(and `typescript-eslint` then pulls in a working version of `eslint`
with it).