Rewrite the programming guide to use the new guestbook (#168)
1. The communicator example is now using a real DBOS Cloud Guestbook
App. In order to use the communicator we also use a key-in-the-config
pattern. I thought that would be a good thing to illustrate.

2. In the last stage, I made the workflow invoke asynchronous; otherwise
some browsers would re-send the request as soon as the app comes back
online and that made the workflow reliability demo hard to follow.

3. I added a small example of reading from the database. I found this
informative (in my own learning journey :)) and the only example that
uses `readOnly` transactions. It also dovetails with 2: now that
workflows are async, you can use this endpoint to check when the
greeting finished.

4. Now that we have writing to the db and reading from the db (3 above)
it was a lot cleaner to use Knex with a struct instead of SQL literal to
write queries. Why not?

5. Made sure all the code examples are linter friendly

6. Language style changes and simplification.

---------

Signed-off-by: Alex Poliakov <alex.poliakov@dbos.dev>
Co-authored-by: Qian Li <qian.li@dbos.dev>
Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>