V2 Docs (#276)
Major doc redo phase 1:
1. Move things around to avoid wall of things on the left
2. Redirects for same
3. Update many sections to use V2 API

There are some noted exceptions / areas that need real work, sometimes
in the SDK. Those will go in a tracking sheet.

Otherwise, it is much easier to review and fix this after it is
merged... as a diff is a bit of a struggle. Lots of follow-on PRs would
be better.