Decorator and Context split (#334)
Split the part of decorators and contexts that are "v1" and are fully
replaced from those that are not.
Remove InitContext, which, while it still works harmlessly, is entirely
optional.
Parameter decorators also work, but there is no reason to know about
them, so they are removed from the doc.

This is just an increment.  Next steps:
Improve some of the product features that are responsible for more of
the documented things we don't like (ORMs, HTTP, etc.)
Do JSDocs, to contribute reference material for a proper reference,
closely related matter that "DBOS Class" doc is too big.
  Drop the olddecorators / oldcontexts pages whenever we are ready